requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.46.0
django-braces==1.15.0
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-haystack==3.1.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.2.0
django-modeltranslation==0.17.3
django-star-ratings==0.9.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pysolr==3.9.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.7
whitenoise==5.3.0
